+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 ## Added
+- Added new `Notification` component for displaying notifications at top of page [#191]
 - Added `UpdateAffiliation` and `AffiliationById` queries [#203]
 
 ## Updated
+- Updated `PlanOverviewPage` component to display a feedback mode notification at top of page when one has status=REQUESTED [#191]
 - Enhanced condition in `RepoSelectorForAnswer` to display the preferred Repositories checkbox [#118]
 - Updated `FeedbackOptions` component to use real data [#203]